Replace option parsing loop with optparse.OptionParser.
This will allow us to reduce doc duplication. It will also make adding
new options easier and ensure consistent documentation, since the
entries for options will be automatically added to the usage statement.

Note that optparse is deprecated in favor of argparse. We use optparse
anyway for backward compatibility with systems running older versions of
Python.