Reformat usage statement to use optparse.
Admittedly, this isn't as much of a simplification as I thought it would
be. For example, we need to include a subclass of a formatter to work
around what it does to newlines. We also change the way that the usage
statement looks; now, options are interleaved into descriptions. I'm not
sure whether I like it yet, but it works.